ml-api/requirements.txt
# Flask backend
Flask>=2.3.0
flask-cors>=4.0.0

# ML
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Optional: transformers for DistilBERT (SMS) - using TF-IDF fallback by default
# transformers>=4.30.0
# torch>=2.0.0

# Image / OCR support
Pillow>=10.0.0
pytesseract>=0.3.10
python-multipart>=0.0.6
werkzeug>=3.0.0

# Gemini AI for human-readable explanations (why legit / why fake)
google-generativeai>=0.8.0
python-dotenv>=1.0.0
